Make tbi.py file for run_nth_year_behresp_model
